perf: <AvatarWithFallback />
improved performance by accepting an optional `width` prop and passing it to the <Image /> component, which resizes the image if the `width` prop is present. Also now accepting the `sizes` prop to improve performance because when the `width` prop is not specified, the component sets the `fill` prop of the <Image /> component to true, which in turn displays a warning if the `sizes` prop is not present.