fix not being able to see question after adding it
the bug was that you could only see the newly added question after you refreshed the page when it got loaded from localStorage. This happened because i used a custom hook `useAddTestFormQuestionsArr`, which was called from multiple components and each of those components received a different instance of the return value of `useFieldArray`. As a mostly sane person, i though it was fine because the `useFieldArray` hook was being passed the same exact `control` option, but apparently all components have to use the same reference to the return value of the `useFieldArray` hook, so now I'm calling this hook in the <AddTestForm /> component and passing its return value down as props. Though this is prop drilling so i might need to learn about zustand a bit to be able to pass this value through a store.